requirements.txt
streamlit
mysql-connector-python
python-dotenv
passlib[bcrypt]
